[CR-1233677] drv:src: Updated dma pad value checks as per AIE4 also.
This patch handles checking before and after pad max values according to
respective device specific values. Added check condition for AIE4 device
as well.

Since the max value is 8bit for AIE4, the data types changed to
u16 considering future requirements and we can have porper check
condition for max values in AIE4 device also. If the data type is not changed then
driver can not check max values for AIE4

+++ driver/src/global/xaiegbl.h
@@ -397,8 +397,8 @@ typedef union {
 } XAie_MultiDimDesc;
 
 typedef struct {
-	u8 Before;
-	u8 After;
+	u16 Before;
+	u16 After;
 } XAie_PadDesc;
 
 typedef struct {
